+++ src/main/java/blackjack/domain/card/Card.java
@@ -1,6 +1,6 @@
 package blackjack.domain.card;
 
-import blackjack.domain.game.Score;
+import blackjack.domain.participant.Score;
 import java.util.Objects;
 
 public class Card {
