+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,8 +1,6 @@
 package blackjack.domain.participant;
 
 import blackjack.domain.card.CardDeck;
-import blackjack.domain.card.Hand;
-import blackjack.domain.game.Score;
 
 public class Dealer {
