+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,8 +1,6 @@
 package blackjack.domain.participant;
 
 import blackjack.domain.card.CardDeck;
-import blackjack.domain.card.Hand;
-import blackjack.domain.game.Score;
 import java.util.Objects;
 
 public class Player {
