+++ src/main/java/blackjack/domain/participant/Score.java
@@ -1,4 +1,4 @@
-package blackjack.domain.game;
+package blackjack.domain.participant;
 
 import java.util.Objects;
