+++ src/main/java/blackjack/view/mapper/DrawDecisionMapper.java
@@ -1,6 +1,6 @@
 package blackjack.view.mapper;
 
-import blackjack.domain.game.DrawDecision;
+import blackjack.domain.participant.DrawDecision;
 import java.util.Arrays;
 
 public enum DrawDecisionMapper {
